use protected property
